Finish conversion to netcdf. Plotting from netcdf files. netcdf files with global grid
